+++ R/pixgallery.R
@@ -1,90 +1,103 @@
-
-#' @title Create an image gallery.
-#' @description Create an image gallery.
-#' @param path [Character Vector] Full paths to images.
-#' @param caption [Character Vector] Captions for the images equal to length of path.
-#' @param caption_valign [Character] Position of the caption. Options are 'none', 'top', 'center', 'bottom' or 'below'.
-#' @param caption_halign [Character] Horizontal justification of the caption. Options are 'left', 'center' or 'right'.
-#' @param link [Logical/Character Vector] TRUE for lightbox, FALSE to disable the lightbox or a character vector of custom URLs equal to length of path.
-#' @param h [Character] Height of the image thumbnails in valid css units.
-#' @param w [Character] Width of the image thumbnails in valid css units.
-#' @param gap [Character] Gap between thumbnails in valid css units.
-#' @param border_radius [Character] Corner radius of image thumbnails in valid css units.
-#' @param layout [Character] Gallery layout type. Options are "grid", "fixed", "mosaic", "masonry", "justified", "elastic" or "rhombus". See details.
-#' @param shuffle [Logical] A logical indicating whether images are randomly shuffled.
-#' @param height [Character] Height of the widget in valid css units.
-#' @param width [Character] Width of the widget in valid css units.
-#' @param elementId [Character] Parent container ID.
+#' @title Create An HTML Image Gallery
+#' @description Create an HTML image gallery
+#' @param path A character vector of full paths to images.
+#' @param caption A character vector of captions for the images (Optional). It must be equal to the length of path.
+#' @param caption_valign A character denoting position of the caption. Options are 'none', 'top', 'center', 'bottom' or 'below'.
+#' @param caption_halign A character denoting horizontal justification of the caption. Options are 'left', 'center' or 'right'.
+#' @param link A logical or character vector. What happens when you click on the thumbnail? TRUE opens up the lightbox, FALSE to disable the lightbox. A character vector of custom URLs equal to length of path.
+#' @param h A character denoting height of the image thumbnails in valid CSS units.
+#' @param w A character denoting width of the image thumbnails in valid CSS units.
+#' @param gap A character denoting spacing between thumbnails in valid CSS units.
+#' @param border_radius A character denoting corner radius of image thumbnails in valid CSS units.
+#' @param layout A character denoting gallery layout type. Options are "grid", "fixed", "mosaic", "masonry", "justified", "elastic" or "rhombus". See details.
+#' @param shuffle A logical indicating whether images are randomly shuffled.
+#' @param height A character denoting height of the widget in valid CSS units.
+#' @param width A character denoting width of the widget in valid CSS units.
+#' @param elementId A character string denoting parent container ID.
+#' @param dim Deprecated. Use 'h' and 'w'.
 #' @param type Deprecated. Use 'layout'.
 #' @importFrom htmlwidgets createWidget sizingPolicy
-#' @details 
-#' grid: Square grid responsive layout. Adjust square size using 'dim'. Width is flexible to fit parent. \cr
-#' fixed: Same as grid but strictly fixed dimensions. \cr
-#' masonry: Column responsive layout with fixed width. Adjust column width using 'dim'. \cr
-#' justified: Similar to masonry but rows have fixed height. Adjust row height using 'dim'. \cr
-#' elastic: Single row layout where images are magnified on hover. \cr
-#' mosaic: Grid layout with mixed sizes. \cr
-#' rhombus: Diamond shaped layout with three columns. Size and number of columns are fixed. Gap can be adjusted.
+#' @details
+#' **grid**: Grid responsive layout. Width is flexible to fit parent. \cr
+#' **fixed**: Same as grid but non-responsive. Strictly fixed dimensions. Defaults to square image. When using `caption_valign` options 'top','bottom' or 'center', note that long captions may overflow. \cr
+#' **mosaic**: Grid layout with mixed sizes. \cr
+#' **masonry**: Column layout where images flow vertically. Image height cannot be set. When using `caption_valign` options 'top','bottom' or 'center', note that long captions may overflow. \cr
+#' **justified**: Similar to masonry but images flow horizontally. Image width cannot be set. When using `caption_valign` options 'top','bottom' or 'center', note that long captions are clipped. Option 'below' may cause images to be stretched if captions are too long. \cr
+#' **elastic**: Single row layout where images are magnified on hover. Image width cannot be set. Captions cannot be displayed on thumbnails. \cr
+#' **rhombus**: Diamond shaped layout with three columns. Image size and number of columns are fixed. Completely non-responsive without breakpoints. Captions cannot be displayed on thumbnails. \cr
+#' 
+#' @examples
+#' library(pixture)
+#' paths <- list.files(path=system.file("extdata/images",package="pixture"),full.names=TRUE)
+#' pixgallery(paths)
+#' 
 #' @export
 #'
 pixgallery <- function(
-  path,
-  caption = NULL,
-  caption_valign = "none",
-  caption_halign = "left",
-  link = TRUE,
-  h = NULL,
-  w = NULL,
-  gap = "5px",
-  border_radius = "0px",
-  layout="grid",
-  shuffle = FALSE,
-  width = "100%",
-  height = "100%",
-  elementId = NULL,
-  type = NULL
-  ) {
-
+    path,
+    caption = NULL,
+    caption_valign = "none",
+    caption_halign = "left",
+    link = TRUE,
+    h = NULL,
+    w = NULL,
+    gap = "5px",
+    border_radius = "0px",
+    layout = "grid",
+    shuffle = FALSE,
+    width = "100%",
+    height = "100%",
+    elementId = NULL,
+    dim = NULL,
+    type = NULL) {
+  
+  # check type
+  if (!is.null(type)) {
+    layout <- type
+    warning("Parameter 'type' is deprecated. Please use 'layout'.")
+  }
+  
+  # check type
+  if (!is.null(dim)) {
+    h <- dim
+    w <- dim
+    warning("Parameter 'dim' is deprecated. Please use 'h' and/or 'w' to set height and width of thumbnails.")
+  }
+  
   # check caption
-  if(!is.null(caption)) {
-    if(length(caption) != length(path)) stop(paste0("Length of 'caption' (", length(caption), ") is not not equal to the length of 'path' (", length(path), "). If 'caption' is used, it must be the same length as 'path'."))
+  if (!is.null(caption)) {
+    if (length(caption) != length(path)) stop(paste0("Length of 'caption' (", length(caption), ") is not not equal to the length of 'path' (", length(path), "). If 'caption' is used, it must be the same length as 'path'."))
   } else {
-    if(!is.null(names(path))) caption <- names(path)
+    if (!is.null(names(path))) caption <- names(path)
   }
   names(path) <- NULL
 
   # check caption position
   caption_valigns <- c("none", "top", "center", "bottom", "below")
-  if (!caption_valign %in% caption_valigns) stop(paste0("Parameter 'caption_valign' must be one of '", paste(caption_valigns, collapse = "', '"), "'."))
+  if (is.null(caption_valign) || !caption_valign %in% caption_valigns) stop(paste0("Parameter 'caption_valign' must be one of '", paste(caption_valigns, collapse = "', '"), "'."))
 
   # check caption justification
   caption_haligns <- c("left", "center", "right")
-  if (!caption_halign %in% caption_haligns) stop(paste0("Parameter 'caption_halign' must be one of '", paste(caption_haligns, collapse = "', '"), "'."))
-
-  # check type
-  if(!is.null(type)){
-    layout <- type
-    warning("Parameter 'type' is deprecated. Please use 'layout'.")
-  }
+  if (is.null(caption_halign) || (!caption_halign %in% caption_haligns)) stop(paste0("Parameter 'caption_halign' must be one of '", paste(caption_haligns, collapse = "', '"), "'."))
 
   # check link
-  if (is.logical(link)) {
-    if (length(link) != 1) {
-      stop("Parameter 'link' must be of length 1 if TRUE or FALSE.")
-    }
-  } else if (is.character(link)) {
+  if (is.null(link) || all(is.na(link))) stop("Parameter 'link' must be a logical or a character vector.")
+  if (is.logical(link) && (length(link) != 1)) stop("Parameter 'link' must be of length 1. TRUE or FALSE.")
+  if (is.character(link)) {
     if (length(link) != length(path)) {
       stop(paste0("Length of 'link' (", length(link), ") is not not equal to the length of 'path' (", length(path), "). If 'link' is a character, it must be the same length as 'path'."))
     }
-  } else {
-    stop("Parameter 'link' must be a logical or a character vector.")
   }
 
+  # check layout
   layouts <- c("grid", "fixed", "mosaic", "elastic", "masonry", "justified", "rhombus")
-  if(!layout %in% layouts) stop(paste0("Parameter 'layout' must be one of '",paste(layouts,collapse="', '"),"'."))
+  if (is.null(layout) || (!layout %in% layouts)) stop(paste0("Parameter 'layout' must be one of '", paste(layouts, collapse = "', '"), "'."))
+
+  # check shuffle
+  if (is.null(shuffle) || is.na(shuffle) || (!is.logical(shuffle))) stop("Parameter 'shuffle' must be a logical. TRUE or FALSE.")
 
   # forward options using x
-  x = list(
+  x <- list(
     path = as.list(path),
     caption = as.list(caption),
     caption_valign = caption_valign,
@@ -100,11 +113,11 @@ pixgallery <- function(
 
   # create widget
   createWidget(
-    name = 'pixgallery',
+    name = "pixgallery",
     x,
     width = width,
     height = height,
-    package = 'pixture',
+    package = "pixture",
     elementId = elementId,
     sizingPolicy = sizingPolicy(
       defaultWidth = "100%",
@@ -133,8 +146,8 @@ pixgallery <- function(
 #' @importFrom htmlwidgets shinyWidgetOutput
 #' @export
 #'
-pixgalleryOutput <- function(outputId, width = '100%', height = '100%'){
-  shinyWidgetOutput(outputId, 'pixgallery', width, height, package = 'pixture')
+pixgalleryOutput <- function(outputId, width = "100%", height = "100%") {
+  shinyWidgetOutput(outputId, "pixgallery", width, height, package = "pixture")
 }
 
 #' @rdname pixgallery-shiny
@@ -142,7 +155,9 @@ pixgalleryOutput <- function(outputId, width = '100%', height = '100%'){
 #' @export
 #'
 renderPixgallery <- function(expr, env = parent.frame(), quoted = FALSE) {
-  if (!quoted) { expr <- substitute(expr) } # force quoted
+  if (!quoted) {
+    expr <- substitute(expr)
+  } # force quoted
   shinyRenderWidget(expr, pixgalleryOutput, env, quoted = TRUE)
 }